requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
beanie==1.26.0
motor==3.4.0
pydantic==2.7.4
pydantic-settings==2.3.4
python-jose[cryptography]==3.3.0
bcrypt>=4.1,<6.0
python-multipart==0.0.9
boto3==1.34.133
cloudinary==1.40.0
email-validator==2.1.1
fastapi-mail==1.4.1
